move messages to messagesSection
